# openapi.yaml
openapi: 3.0.3
info:
  title: AI Compliance Monitor API
  description: Comprehensive AI compliance monitoring system for MedTech with real-time safety checking, audit reporting, and regulatory compliance dashboard
  version: 1.0.0
  contact:
    name: AI Compliance Monitor
    url: https://your-domain.com
servers:
  - url: https://your-domain.com/api
    description: Production API Server

paths:
  /check-output:
    post:
      summary: Check AI output for compliance violations
      description: Analyze AI model output for HIPAA, GDPR, and other regulatory compliance issues
      operationId: checkOutput
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - input
                - output
                - modelName
              properties:
                input:
                  type: string
                  description: The input text provided to the AI model
                  example: "Patient John Doe has diabetes"
                output:
                  type: string
                  description: The output generated by the AI model
                  example: "Based on the patient's condition, recommend insulin therapy"
                modelName:
                  type: string
                  description: Name of the AI model that generated the output
                  example: "MedicalAI-v2"
                metadata:
                  type: object
                  description: Additional metadata about the AI output
                  additionalProperties: true
      responses:
        '200':
          description: Compliance check completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique identifier for the compliance check
                  status:
                    type: string
                    enum: [safe, flagged, blocked]
                    description: Compliance status of the AI output
                  riskScore:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                    description: Risk score from 0 (safe) to 1 (high risk)
                  reasons:
                    type: array
                    items:
                      type: string
                    description: List of reasons why the output was flagged or blocked
                  timestamp:
                    type: string
                    format: date-time
                    description: When the compliance check was performed
        '400':
          description: Invalid request data
        '500':
          description: Internal server error

  /scan-training-data:
    post:
      summary: Scan training data for compliance issues
      description: Upload and analyze training datasets for privacy risks, bias, and compliance violations
      operationId: scanTrainingData
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Training data file (CSV, JSON, or XLSX)
      responses:
        '200':
          description: Training data scan completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique identifier for the scan
                  fileName:
                    type: string
                    description: Name of the uploaded file
                  totalRows:
                    type: integer
                    description: Total number of rows in the dataset
                  flaggedRows:
                    type: integer
                    description: Number of rows flagged for compliance issues
                  privacyRisks:
                    type: integer
                    description: Number of privacy risks detected
                  biasFlags:
                    type: integer
                    description: Number of bias flags identified
                  missingDocs:
                    type: integer
                    description: Number of missing documentation issues
                  status:
                    type: string
                    enum: [completed, processing, failed]
                    description: Status of the scan
                  timestamp:
                    type: string
                    format: date-time
                    description: When the scan was completed
        '400':
          description: Invalid file or request
        '500':
          description: Internal server error

  /report/{id}:
    get:
      summary: Get audit report by ID
      description: Retrieve a specific compliance audit report
      operationId: getAuditReport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Unique identifier of the audit report
      responses:
        '200':
          description: Audit report retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditReport'
        '404':
          description: Report not found
        '500':
          description: Internal server error

  /metrics:
    get:
      summary: Get real-time compliance metrics
      description: Retrieve current compliance statistics and metrics
      operationId: getMetrics
      responses:
        '200':
          description: Metrics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  complianceScore:
                    type: number
                    format: float
                    description: Overall compliance score percentage
                  flaggedOutputs24h:
                    type: integer
                    description: Number of flagged outputs in the last 24 hours
                  modelDriftPercent:
                    type: number
                    format: float
                    description: Model drift percentage
                  activeAudits:
                    type: integer
                    description: Number of active audit processes
                  hipaaCompliance:
                    type: boolean
                    description: HIPAA compliance status
                  gdprCompliance:
                    type: boolean
                    description: GDPR compliance status
                  totalChecks24h:
                    type: integer
                    description: Total compliance checks in the last 24 hours
        '500':
          description: Internal server error

  /logs:
    get:
      summary: Get AI output check logs
      description: Retrieve list of recent AI output compliance checks
      operationId: getLogs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Maximum number of logs to return
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiOutputCheck'
        '500':
          description: Internal server error

  /generate-report:
    post:
      summary: Generate new audit report
      description: Create a new compliance audit report for specified framework
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reportType
                - complianceFramework
              properties:
                reportType:
                  type: string
                  enum: [compliance-audit, risk-assessment, privacy-impact, security-review]
                  description: Type of report to generate
                complianceFramework:
                  type: string
                  enum: [HIPAA, GDPR, FDA, SOC2]
                  description: Regulatory framework for the report
                description:
                  type: string
                  description: Optional description for the report
      responses:
        '200':
          description: Report generation initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditReport'
        '400':
          description: Invalid request data
        '500':
          description: Internal server error

  /alerts:
    get:
      summary: Get compliance alerts
      description: Retrieve active compliance alerts and notifications
      operationId: getAlerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, all]
          description: Filter alerts by status
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Maximum number of alerts to return
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplianceAlert'
        '500':
          description: Internal server error

  /drift-data:
    get:
      summary: Get model drift data
      description: Retrieve model performance drift information
      operationId: getDriftData
      parameters:
        - name: model
          in: query
          schema:
            type: string
          description: Filter by specific model name
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Number of days of historical data to retrieve
      responses:
        '200':
          description: Drift data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDriftData'
        '500':
          description: Internal server error

  /bias-results:
    get:
      summary: Get bias detection results
      description: Retrieve bias detection analysis results across demographic groups
      operationId: getBiasResults
      parameters:
        - name: model
          in: query
          schema:
            type: string
          description: Filter by specific model name
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Number of days of historical data to retrieve
      responses:
        '200':
          description: Bias results retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BiasDetectionResult'
        '500':
          description: Internal server error

components:
  schemas:
    AiOutputCheck:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        timestamp:
          type: string
          format: date-time
          description: When the check was performed
        modelName:
          type: string
          description: Name of the AI model
        inputData:
          type: string
          description: Input text provided to the model
        outputData:
          type: string
          description: Output generated by the model
        status:
          type: string
          enum: [safe, flagged, blocked]
          description: Compliance status
        riskScore:
          type: number
          format: float
          description: Risk score from 0 to 1
        reasons:
          type: array
          items:
            type: string
          description: Reasons for flagging
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata

    AuditReport:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        timestamp:
          type: string
          format: date-time
          description: When the report was generated
        reportType:
          type: string
          description: Type of audit report
        title:
          type: string
          description: Report title
        description:
          type: string
          description: Report description
        status:
          type: string
          enum: [generated, pending, failed]
          description: Report generation status
        complianceFramework:
          type: string
          description: Regulatory framework (HIPAA, GDPR, etc.)
        findings:
          type: object
          additionalProperties: true
          description: Audit findings and results
        recommendations:
          type: array
          items:
            type: string
          description: Recommended actions
        pdfPath:
          type: string
          description: Path to PDF version of report

    ComplianceAlert:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        timestamp:
          type: string
          format: date-time
          description: When the alert was created
        alertType:
          type: string
          description: Type of compliance alert
        severity:
          type: string
          enum: [low, medium, high, critical]
          description: Alert severity level
        title:
          type: string
          description: Alert title
        description:
          type: string
          description: Alert description
        modelName:
          type: string
          description: Associated model name (if applicable)
        status:
          type: string
          enum: [active, resolved]
          description: Alert status
        resolvedAt:
          type: string
          format: date-time
          description: When the alert was resolved

    ModelDriftData:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        timestamp:
          type: string
          format: date-time
          description: When the drift measurement was taken
        modelName:
          type: string
          description: Name of the model
        driftScore:
          type: number
          format: float
          description: Drift score percentage
        baselineAccuracy:
          type: number
          format: float
          description: Baseline accuracy percentage
        currentAccuracy:
          type: number
          format: float
          description: Current accuracy percentage
        dataDistributionShift:
          type: number
          format: float
          description: Data distribution shift measurement
        performanceMetrics:
          type: object
          additionalProperties: true
          description: Additional performance metrics

    BiasDetectionResult:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
        timestamp:
          type: string
          format: date-time
          description: When the bias analysis was performed
        modelName:
          type: string
          description: Name of the model
        demographicGroup:
          type: string
          description: Demographic group analyzed
        biasScore:
          type: number
          format: float
          description: Bias score for the demographic group
        fairnessMetrics:
          type: object
          additionalProperties: true
          description: Fairness metrics and measurements
        mitigationSuggestions:
          type: array
          items:
            type: string
          description: Suggested bias mitigation strategies

tags:
  - name: compliance
    description: AI compliance checking operations
  - name: reports
    description: Audit report management
  - name: monitoring
    description: Real-time monitoring and metrics
  - name: analysis
    description: Data analysis and scanning