# server/routes.ts
import type { Express } from "express";
import { createServer, type Server } from "http";
import { storage } from "./storage";
import { z } from "zod";
import { 
  insertAiOutputCheckSchema, insertTrainingDataScanSchema, 
  insertAuditReportSchema, insertComplianceMetricsSchema,
  insertModelDriftDataSchema, insertBiasDetectionResultSchema,
  insertComplianceAlertSchema
} from "@shared/schema";
import multer from "multer";
import path from "path";
import fs from "fs";

const upload = multer({ dest: "uploads/" });

// Validation schemas for API endpoints
const checkOutputSchema = z.object({
  input: z.string(),
  output: z.string(),
  modelName: z.string(),
  metadata: z.record(z.any()).optional().default({}),
});

const scanTrainingDataSchema = z.object({
  fileName: z.string(),
  fileSize: z.number(),
  totalRows: z.number(),
  scanResults: z.record(z.any()),
});

export async function registerRoutes(app: Express): Promise<Server> {
  
  // CORS middleware for external API access
  app.use((req, res, next) => {
    res.header('Access-Control-Allow-Origin', '*');
    res.header('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
    res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Authorization');
    if (req.method === 'OPTIONS') {
      res.sendStatus(200);
    } else {
      next();
    }
  });
  
  // POST /api/check-output - AI output safety checking
  app.post("/api/check-output", async (req, res) => {
    try {
      const { input, output, modelName, metadata } = checkOutputSchema.parse(req.body);
      
      // Perform compliance checking logic
      const riskScore = calculateRiskScore(input, output);
      const status = determineStatus(riskScore);
      const reasons = generateReasons(input, output, riskScore);
      
      const check = await storage.createAiOutputCheck({
        modelName,
        inputData: input,
        outputData: output,
        status,
        riskScore,
        reasons,
        metadata,
      });

      // Update compliance metrics
      await updateComplianceMetrics();
      
      res.json({
        id: check.id,
        status,
        riskScore,
        reasons,
        timestamp: check.timestamp,
      });
    } catch (error) {
      console.error("Error checking AI output:", error);
      res.status(500).json({ error: "Failed to check AI output" });
    }
  });

  // POST /api/scan-training-data - Training data compliance scan
  app.post("/api/scan-training-data", upload.single("file"), async (req, res) => {
    try {
      if (!req.file) {
        return res.status(400).json({ error: "No file uploaded" });
      }

      const filePath = req.file.path;
      const fileName = req.file.originalname;
      const fileSize = req.file.size;
      
      // Perform training data analysis
      const analysisResults = await analyzeTrainingData(filePath);
      
      const scan = await storage.createTrainingDataScan({
        fileName,
        fileSize,
        totalRows: analysisResults.totalRows,
        flaggedRows: analysisResults.flaggedRows,
        privacyRisks: analysisResults.privacyRisks,
        biasFlags: analysisResults.biasFlags,
        missingDocs: analysisResults.missingDocs,
        scanResults: analysisResults as Record<string, any>,
        status: "completed",
      });

      // Clean up uploaded file
      fs.unlinkSync(filePath);
      
      res.json({
        id: scan.id,
        fileName,
        totalRows: scan.totalRows,
        flaggedRows: scan.flaggedRows,
        privacyRisks: scan.privacyRisks,
        biasFlags: scan.biasFlags,
        missingDocs: scan.missingDocs,
        status: scan.status,
        timestamp: scan.timestamp,
      });
    } catch (error) {
      console.error("Error scanning training data:", error);
      res.status(500).json({ error: "Failed to scan training data" });
    }
  });

  // GET /api/report/:id - Get audit report
  app.get("/api/report/:id", async (req, res) => {
    try {
      const id = parseInt(req.params.id);
      const report = await storage.getAuditReport(id);
      
      if (!report) {
        return res.status(404).json({ error: "Report not found" });
      }

      res.json(report);
    } catch (error) {
      console.error("Error fetching report:", error);
      res.status(500).json({ error: "Failed to fetch report" });
    }
  });

  // GET /api/metrics - Real-time compliance metrics
  app.get("/api/metrics", async (req, res) => {
    try {
      const metrics = await storage.getLatestComplianceMetrics();
      
      if (!metrics) {
        // Generate initial metrics if none exist
        const initialMetrics = await generateInitialMetrics();
        return res.json(initialMetrics);
      }

      res.json(metrics);
    } catch (error) {
      console.error("Error fetching metrics:", error);
      res.status(500).json({ error: "Failed to fetch metrics" });
    }
  });

  // GET /api/metrics/history - Compliance metrics history
  app.get("/api/metrics/history", async (req, res) => {
    try {
      const days = parseInt(req.query.days as string) || 7;
      const history = await storage.getComplianceMetricsHistory(days);
      res.json(history);
    } catch (error) {
      console.error("Error fetching metrics history:", error);
      res.status(500).json({ error: "Failed to fetch metrics history" });
    }
  });

  // GET /api/logs - AI output check logs
  app.get("/api/logs", async (req, res) => {
    try {
      const limit = parseInt(req.query.limit as string) || 50;
      const logs = await storage.getAiOutputChecks(limit);
      res.json(logs);
    } catch (error) {
      console.error("Error fetching logs:", error);
      res.status(500).json({ error: "Failed to fetch logs" });
    }
  });

  // GET /api/training-scans - Training data scan results
  app.get("/api/training-scans", async (req, res) => {
    try {
      const limit = parseInt(req.query.limit as string) || 50;
      const scans = await storage.getTrainingDataScans(limit);
      res.json(scans);
    } catch (error) {
      console.error("Error fetching training scans:", error);
      res.status(500).json({ error: "Failed to fetch training scans" });
    }
  });

  // GET /api/drift-data - Model drift data
  app.get("/api/drift-data", async (req, res) => {
    try {
      const modelName = req.query.model as string;
      const days = parseInt(req.query.days as string) || 30;
      const driftData = await storage.getModelDriftData(modelName, days);
      res.json(driftData);
    } catch (error) {
      console.error("Error fetching drift data:", error);
      res.status(500).json({ error: "Failed to fetch drift data" });
    }
  });

  // GET /api/bias-results - Bias detection results
  app.get("/api/bias-results", async (req, res) => {
    try {
      const modelName = req.query.model as string;
      const days = parseInt(req.query.days as string) || 30;
      const biasResults = await storage.getBiasDetectionResults(modelName, days);
      res.json(biasResults);
    } catch (error) {
      console.error("Error fetching bias results:", error);
      res.status(500).json({ error: "Failed to fetch bias results" });
    }
  });

  // GET /api/alerts - Compliance alerts
  app.get("/api/alerts", async (req, res) => {
    try {
      const status = req.query.status as string;
      const limit = parseInt(req.query.limit as string) || 50;
      const alerts = await storage.getComplianceAlerts(status, limit);
      res.json(alerts);
    } catch (error) {
      console.error("Error fetching alerts:", error);
      res.status(500).json({ error: "Failed to fetch alerts" });
    }
  });

  // GET /api/audit-reports - Audit reports list
  app.get("/api/audit-reports", async (req, res) => {
    try {
      const limit = parseInt(req.query.limit as string) || 50;
      const reports = await storage.getAuditReports(limit);
      res.json(reports);
    } catch (error) {
      console.error("Error fetching audit reports:", error);
      res.status(500).json({ error: "Failed to fetch audit reports" });
    }
  });

  // Serve OpenAPI specification
  app.get("/openapi.yaml", (req, res) => {
    res.sendFile("/openapi.yaml", { root: process.cwd() });
  });

  // Serve plugin manifest
  app.get("/.well-known/ai-plugin.json", (req, res) => {
    res.sendFile("/ai-plugin.json", { root: process.cwd() });
  });

  // POST /api/generate-report - Generate new audit report
  app.post("/api/generate-report", async (req, res) => {
    try {
      const { reportType, complianceFramework, description } = req.body;
      
      const findings = await generateAuditFindings(complianceFramework);
      const recommendations = generateRecommendations(findings);
      
      const report = await storage.createAuditReport({
        reportType,
        title: `${complianceFramework} Compliance Report - ${new Date().toLocaleDateString()}`,
        description,
        status: "generated",
        complianceFramework,
        findings: findings as Record<string, any>,
        recommendations,
      });

      res.json(report);
    } catch (error) {
      console.error("Error generating report:", error);
      res.status(500).json({ error: "Failed to generate report" });
    }
  });

  const httpServer = createServer(app);
  return httpServer;
}

// Helper functions for compliance logic
function calculateRiskScore(input: string, output: string): number {
  let score = 0;
  
  // Check for PHI patterns
  const phiPatterns = [
    /\b\d{3}-\d{2}-\d{4}\b/, // SSN
    /\b\d{10,11}\b/, // Phone numbers
    /\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b/, // Email
    /\b\d{1,2}\/\d{1,2}\/\d{4}\b/, // Dates
  ];
  
  for (const pattern of phiPatterns) {
    if (pattern.test(input) || pattern.test(output)) {
      score += 0.3;
    }
  }
  
  // Check for sensitive medical terms
  const sensitiveTerms = [
    "diagnosis", "treatment", "medication", "prescription", "medical record",
    "patient", "symptom", "condition", "therapy", "surgery"
  ];
  
  const text = (input + " " + output).toLowerCase();
  for (const term of sensitiveTerms) {
    if (text.includes(term)) {
      score += 0.1;
    }
  }
  
  return Math.min(score, 1.0);
}

function determineStatus(riskScore: number): string {
  if (riskScore >= 0.8) return "blocked";
  if (riskScore >= 0.5) return "flagged";
  return "safe";
}

function generateReasons(input: string, output: string, riskScore: number): string[] {
  const reasons: string[] = [];
  
  if (riskScore >= 0.5) {
    if (/\b\d{3}-\d{2}-\d{4}\b/.test(input + output)) {
      reasons.push("Potential SSN detected");
    }
    if (/\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b/.test(input + output)) {
      reasons.push("Email address detected");
    }
    if ((input + output).toLowerCase().includes("patient")) {
      reasons.push("Patient information referenced");
    }
  }
  
  return reasons;
}

async function analyzeTrainingData(filePath: string) {
  // Simulate training data analysis
  const stats = fs.statSync(filePath);
  const totalRows = Math.floor(Math.random() * 10000) + 1000;
  const flaggedRows = Math.floor(totalRows * 0.05);
  const privacyRisks = Math.floor(flaggedRows * 0.3);
  const biasFlags = Math.floor(flaggedRows * 0.4);
  const missingDocs = Math.floor(flaggedRows * 0.3);
  
  return {
    totalRows,
    flaggedRows,
    privacyRisks,
    biasFlags,
    missingDocs,
    fileSize: stats.size,
    analysisComplete: true,
  };
}

async function updateComplianceMetrics() {
  const recentChecks = await storage.getRecentAiOutputChecks(24);
  const flaggedCount = recentChecks.filter(check => check.status === "flagged" || check.status === "blocked").length;
  const totalChecks = recentChecks.length;
  
  const complianceScore = totalChecks > 0 ? ((totalChecks - flaggedCount) / totalChecks) * 100 : 100;
  
  await storage.createComplianceMetrics({
    complianceScore,
    flaggedOutputs24h: flaggedCount,
    modelDriftPercent: Math.random() * 3, // Simulated drift
    activeAudits: 8,
    hipaaCompliance: true,
    gdprCompliance: true,
    totalChecks24h: totalChecks,
  });
}

async function generateInitialMetrics() {
  const metrics = {
    complianceScore: 98.7,
    flaggedOutputs24h: 127,
    modelDriftPercent: 2.3,
    activeAudits: 8,
    hipaaCompliance: true,
    gdprCompliance: true,
    totalChecks24h: 2847,
    timestamp: new Date(),
  };
  
  await storage.createComplianceMetrics(metrics);
  return metrics;
}

async function generateAuditFindings(framework: string) {
  const findings = {
    summary: `${framework} compliance audit completed`,
    issues: [],
    recommendations: [],
    overallScore: Math.floor(Math.random() * 20) + 80,
  };
  
  return findings;
}

function generateRecommendations(findings: any): string[] {
  return [
    "Implement additional access controls",
    "Update privacy notices",
    "Enhance data encryption methods",
    "Conduct regular compliance training",
  ];
}